+++ package.json
@@ -23,6 +23,7 @@
     "react": "^18",
     "react-dom": "^18",
     "react-hot-toast": "^2.4.1",
+    "stripe": "^17.2.1",
     "zod": "^3.23.8"
   },
   "devDependencies": {
