+++ src/app/api/profile/route.ts
@@ -13,6 +13,10 @@ export async function PUT(req: Request) {
   const body = await req.json();
   const userInfoSchema = z
     .object({
+      name: z
+        .string({ message: 'Name should be a string' })
+        .min(1, 'Name is too short')
+        .optional(),
       streetAddress: z
         .string({ message: 'Street address should be a string' })
         .min(1, 'Street address is too short')
@@ -43,8 +47,13 @@ export async function PUT(req: Request) {
   if (!validationResult.success) {
     return Response.json({ error: validationResult.error.issues });
   }
-
-  const email = await getUserEmail();
+  let email;
+  if (!body?._id) {
+    email = await getUserEmail();
+  } else {
+    email = await User.findOne({ _id: body._id });
+    email = email?.email;
+  }
 
   if (!process.env.MONGODB_URI || !process.env.MONGODB_DB) {
     throw new Error('Missing env variables: "MONGODB_URI" Or "MONGODB_DB"');
@@ -52,11 +61,16 @@ export async function PUT(req: Request) {
   await mongoose.connect(process.env.MONGODB_URI, {
     dbName: process.env.MONGODB_DB,
   });
-  const result = await UserInfo.updateOne({ email }, validationResult.data, {
-    upsert: true,
-  });
+  await User.updateOne({ email }, { name: validationResult.data.name });
+  const result = await UserInfo.findOneAndUpdate(
+    { email },
+    validationResult.data,
+    {
+      upsert: true,
+    },
+  );
 
-  if (!result.matchedCount && !result.upsertedCount) {
+  if (!result) {
     return Response.json(
       {
         error: [
@@ -71,10 +85,14 @@ export async function PUT(req: Request) {
   return Response.json({ message: 'User infos has been updated successfully' });
 }
 
-export async function GET() {
-  const email = await getUserEmail();
-
-  if (!email) {
+export async function GET(req: Request) {
+  const url = new URL(req.url);
+  const _id = url.searchParams.get('_id') as string | null;
+  let email;
+  if (!_id) {
+    email = await getUserEmail();
+  }
+  if (!email && !_id) {
     return Response.json({ error: [{ message: 'User email not found' }] });
   }
 
@@ -84,6 +102,11 @@ export async function GET() {
   await mongoose.connect(process.env.MONGODB_URI, {
     dbName: process.env.MONGODB_DB,
   });
-  const user = await getUsers({ email });
+  let user: (typeof User)[] | null = null;
+  if (_id) {
+    user = await getUsers({ _id });
+  } else {
+    user = await getUsers({ email });
+  }
   return Response.json(user);
 }
