+++ src/app/categories/page.tsx
@@ -9,7 +9,7 @@ const CategoriesPage: React.FC = () => {
   const [categoryName, setCategoryName] = useState('');
   const { loading: profileLoading, data: profileData } = useProfile();
   const [editedCategory, setEditedCategory] = useState(null);
-  
+
   const getAllCategories = () => {
     fetch('/api/categories').then((res) => {
       res.json().then((categories) => {
@@ -17,11 +17,11 @@ const CategoriesPage: React.FC = () => {
       });
     });
   };
-  
+
   useEffect(() => {
     getAllCategories();
   }, []);
-  
+
   async function handleCategorySubmit(ev) {
     ev.preventDefault();
     const creationPromise = new Promise(async (resolve, reject) => {
@@ -39,13 +39,13 @@ const CategoriesPage: React.FC = () => {
     });
     await toast.promise(creationPromise, {
       loading: editedCategory
-      ? 'Updating category...'
-      : 'Creating your new category...',
+        ? 'Updating category...'
+        : 'Creating your new category...',
       success: editedCategory ? 'Category updated' : 'Category created',
       error: 'Error, sorry...',
     });
   }
-  
+
   const handleDeleteClick = async (_id) => {
     const deletePromise = new Promise(async (resolve, reject) => {
       const response = await fetch('/api/categories?_id=' + _id, {
@@ -53,23 +53,23 @@ const CategoriesPage: React.FC = () => {
       });
       response.ok ? resolve(null) : reject();
     });
-    
+
     await toast.promise(deletePromise, {
       loading: 'Deleting...',
       success: 'Deleted',
       error: 'Error',
     });
-    
+
     getAllCategories();
   };
 
-    if (profileLoading) {
-      return 'Loading user info...';
-    }
-  
-    if (!profileData?.userInfos?.admin) {
-      return 'Not an admin';
-    }
+  if (profileLoading) {
+    return 'Loading user info...';
+  }
+
+  if (!profileData?.userInfos?.admin) {
+    return 'Not an admin';
+  }
 
   return (
     <section className='mt-8 max-w-lg mx-auto'>
