+++ src/app/menu-items/new/page.tsx
@@ -12,7 +12,7 @@ import toast from 'react-hot-toast';
 const NewMenuItemPage = () => {
   const [redirectToItems, setRedirectToItems] = useState(false);
   const { loading: profileLoading, data: profileData } = useProfile();
-  
+
   async function handleFormSubmit(ev, data) {
     ev.preventDefault();
     const savingPromise = new Promise(async (resolve, reject) => {
@@ -23,27 +23,27 @@ const NewMenuItemPage = () => {
       });
       response.ok ? resolve(null) : reject();
     });
-    
+
     await toast.promise(savingPromise, {
       loading: 'Saving...',
       success: 'Saved',
       error: 'Error',
     });
-    
+
     setRedirectToItems(true);
   }
 
-    if (redirectToItems) {
-      return redirect('/menu-items');
-    }
+  if (redirectToItems) {
+    return redirect('/menu-items');
+  }
 
-    if (profileLoading) {
-      return 'Loading user info...';
-    }
-  
-    if (!profileData?.userInfos?.admin) {
-      return 'Not an admin';
-    }
+  if (profileLoading) {
+    return 'Loading user info...';
+  }
+
+  if (!profileData?.userInfos?.admin) {
+    return 'Not an admin';
+  }
 
   return (
     <section className='mt-8'>
@@ -63,6 +63,6 @@ const NewMenuItemPage = () => {
       />
     </section>
   );
-}
+};
 
 export default NewMenuItemPage;
