+++ src/app/users/page.tsx
@@ -8,13 +8,13 @@ export default function UsersPage() {
   const [users, setUsers] = useState([]);
   const { loading: profileLoading, data: profileData } = useProfile();
 
-    useEffect(() => {
-      fetch('/api/users').then((response) => {
-        response.json().then((users) => {
-          setUsers(users);
-        });
+  useEffect(() => {
+    fetch('/api/users').then((response) => {
+      response.json().then((users) => {
+        setUsers(users);
       });
-    }, []);
+    });
+  }, []);
 
   if (profileLoading) {
     return 'Loading user info...';
